clio info

+++ model/client/server/server_info_response.go
@@ -72,7 +72,7 @@ type ServerLedgerInfo struct {
 	Age            uint          `json:"age"`
 	BaseFeeXRP     float32       `json:"base_fee_xrp"`
 	Hash           types.Hash256 `json:"hash"`
-	ReserveBaseXRP uint          `json:"reserve_base_xrp"`
-	ReserveIncXRP  uint          `json:"reserve_inc_xrp"`
+	ReserveBaseXRP float32       `json:"reserve_base_xrp"`
+	ReserveIncXRP  float32       `json:"reserve_inc_xrp"`
 	Seq            uint          `json:"seq"`
 }
